Починены отрисовка осей, график Si(x), точность обоих синусов. Переделан интерфейс для ввода диапазонов и точности. Лаба готова

+++ LAB №8/integralsinus.h
@@ -10,15 +10,15 @@ public:
     IntegralSinus(unsigned precision)
         : Function<T>(precision, generateIntegrals(precision)) {}
 
-    Array<double> generateIntegrals(int precision) {
-        Array<double> integrals;
+    Array<long double> generateIntegrals(int precision) {
+        Array<long double> integrals;
 
         // Расчет коэффициентов для ряда интегрального синуса
         for (int i = 0; i < precision; ++i) {
             if (i % 2 == 0) {
                 integrals.GetArray().push_back(0); // Члены ряда с четными степенями отсутствуют
             } else {
-                double coefficient = std::pow(-1, (i - 1) / 2) / i;
+                long double coefficient = std::pow(-1, (i - 1) / 2) / i;
                 integrals.GetArray().push_back(coefficient);
             }
         }
